YADR: Python Blender Distributed Renderer
This is a distributed renderer which works in a client/slave mode.
Multiple slaves can run on the renderer nodes. They connect to the
master. The master opens a blender file and reads the animation
then invokes each slave to render a frame.

The assumption is that the master and the slaves all have access
to the same directory structure, shared over the network.

A new C version of the application is being developed which will
provide more extensive and flexible features.

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>